refactor: auth edge and auth fies
